+++ src/components/pages/signIn.tsx
@@ -5,12 +5,11 @@ import backofficeMain from '@/../public/assets/icons/dashboard.svg';
 import permission from '@/../public/assets/icons/permission-rtan.svg';
 import rtan from '@/../public/assets/icons/sign-rtan.svg';
 import write from '@/../public/assets/icons/write-rtan.svg';
+import { postSignIn } from '@/api/authApi';
 import { useModalContext } from '@/context/modal.context';
 import { signupModalType } from '@/types/signup.types';
-import { client } from '@/utils/clientAPI';
 import { loginSchema } from '@/validators/auth/login.validator';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { setCookie } from 'cookies-next';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 import { useForm } from 'react-hook-form';
@@ -43,19 +42,9 @@ const SignIn = () => {
 
   const router = useRouter();
   const onSubmit = async (data: z.infer<typeof loginSchema>) => {
-    try {
-      const res = await client.post('/api/v1/users/login', {
-        username: data.username,
-        password: data.password,
-      });
-      const accessToken = res.headers.authorization.replace('Bearer ', '');
-      if (res?.status === 200) {
-        setCookie('AT', accessToken);
-        router.push('/');
-      }
-    } catch (error) {
-      console.log(error, '로그인 오류');
-    }
+    const response = await postSignIn(data);
+
+    console.log({ response });
   };
 
   return (
